renaming & moving types
